backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
django-filter>=24.3,<24.4
Pillow>=10.4,<10.5
psycopg2-binary>=2.9,<2.10
python-decouple>=3.8,<3.9
argon2-cffi>=23.1,<23.2
